+++ EssentialFeed/EssentialFeed/FeedCache/Infrastructure/CoreData/CoreDataFeedStore+FeedStore.swift
@@ -1,0 +1,45 @@
+//
+//  CoreDataFeedStore+FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Nikhil Menon on 5/6/24.
+//
+
+import Foundation
+
+extension CoreDataFeedStore: FeedStore {
+    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        perform { context in
+            completion(
+                Result {
+                    try ManagedCache.find(in: context).map(context.delete).map(context.save)
+                }
+            )
+        }
+    }
+    
+    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        perform { context in
+            completion(
+                Result {
+                    let managedCache = try ManagedCache.newUniqueInstance(in: context)
+                    managedCache.timestamp = timestamp
+                    managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
+                    try context.save()
+                }
+            )
+        }
+    }
+    
+    public func retrieve(completion: @escaping RetrievalCompletion) {
+        perform { context in
+            completion(
+                Result {
+                    try ManagedCache.find(in: context).map {
+                        CachedFeed(feed: $0.localFeed, timestamp: $0.timestamp)
+                    }
+                }
+            )
+        }
+    }
+}
