+++ EssentialFeed/EssentialFeed/FeedCache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -7,7 +7,7 @@
 
 import CoreData
 
-public final class CoreDataFeedStore: FeedStore {
+public final class CoreDataFeedStore {
     
     private let container: NSPersistentContainer
     private let context: NSManagedObjectContext
@@ -21,41 +21,6 @@ public final class CoreDataFeedStore: FeedStore {
         let context = self.context
         context.perform {
             action(context)
-        }
-    }
-    
-    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        perform { context in
-            completion(
-                Result {
-                    try ManagedCache.find(in: context).map(context.delete).map(context.save)
-                }
-            )
-        }
-    }
-    
-    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        perform { context in
-            completion(
-                Result {
-                    let managedCache = try ManagedCache.newUniqueInstance(in: context)
-                    managedCache.timestamp = timestamp
-                    managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
-                    try context.save()
-                }
-            )
-        }
-    }
-    
-    public func retrieve(completion: @escaping RetrievalCompletion) {
-        perform { context in
-            completion(
-                Result {
-                    try ManagedCache.find(in: context).map {
-                        CachedFeed(feed: $0.localFeed, timestamp: $0.timestamp)
-                    }
-                }
-            )
         }
     }
 }
